+++ modules/perception/fusion/common/camera_util.cc
@@ -21,6 +21,7 @@
 #include "modules/perception/fusion/common/camera_util.h"
 
 #include <limits>
+#include <memory>
 
 namespace apollo {
 namespace perception {
@@ -115,7 +116,7 @@ float ObjectInCameraView(SensorObjectConstPtr sensor_object,
     // use point cloud
     int point_num = cloud.size();
     int in_view_point_num = 0;
-    for (size_t i = 0; i < point_num; ++i) {
+    for (int i = 0; i < point_num; ++i) {
       const auto& pt = cloud.at(i);
       Eigen::Vector3d pt3d(pt.x + offset[0], pt.y + offset[1],
                            pt.z + offset[2]);
@@ -185,9 +186,8 @@ float ObjectInCameraView(SensorObjectConstPtr sensor_object,
   const double dist_slope = 0.25;
   auto sigmoid_like_fun = [max_dist, dist_slope](double obj_dist) {
     double x = obj_dist - max_dist;
-    return 0.5 -
-           0.5 * x * dist_slope /
-               std::sqrt(1 + x * x * dist_slope * dist_slope);
+    return 0.5 - 0.5 * x * dist_slope /
+                     std::sqrt(1 + x * x * dist_slope * dist_slope);
   };
   Eigen::Vector4d center3d_local = world2sensor_pose * center.homogeneous();
   double dist_to_camera = center3d_local.z();
